requirements.txt
git+https://github.com/Pycord-development/pycord.git
python-dotenv
asyncio
pymongo